-Refactored openChat and OpenChatInputBoundary into new package structure
